refactor(backend): move location of entity
